requirements.txt
tensorflow>=2.4.0
dm-sonnet>=2.0.0
absl-py>=1.0.0
numpy>=1.18.0
matplotlib>=3.1.0